feat. subdomain login시 query params 안거치도록

+++ src/templates/Account/Login/index.tsx
@@ -74,38 +74,38 @@ export function LoginTemplate({ tenantName = '', title = '', onSubmitLogin }: Lo
     if (isSuccess) {
       onSubmitLogin();
 
-      const authStore = localStorage.getItem('auth-store');
-      if (authStore) {
-        try {
-          const json = JSON.parse(authStore);
-          if (json && json.state) {
-            const jsonString = JSON.stringify(json.state);
-            // 1. Base64 인코딩 (Node.js 환경에서는 Buffer를 사용)
-            const encodedJson = Buffer.from(jsonString).toString('base64');
-
-            // Continue with your logic here
-            deleteCookie('access_token');
-            localStorage.removeItem('auth-store');
-            localStorage.removeItem('app-storage');
-
-            console.log('loginData', loginData?.redirections?.home_url + `?accessToken=${loginData?.access_token}`);
-            if (username == 're4@naver.com' || username === 're3@naver.com') {
-              location.href = 'http://devus.localhost:3001' + `?authStore=${encodedJson}`;
+      console.log(loginData);
+      if (!loginData?.tenant_uri?.includes(tenantName)) {
+        const authStore = localStorage.getItem('auth-store');
+        if (authStore) {
+          try {
+            const json = JSON.parse(authStore);
+            if (json && json.state) {
+              const jsonString = JSON.stringify(json.state);
+              // 1. Base64 인코딩 (Node.js 환경에서는 Buffer를 사용)
+              const encodedJson = Buffer.from(jsonString).toString('base64');
+              // Continue with your logic here
+              deleteCookie('access_token');
+              localStorage.removeItem('auth-store');
+              localStorage.removeItem('app-storage');
+              console.log('loginData', loginData?.redirections?.home_url + `?accessToken=${loginData?.access_token}`);
+              if (username == 're4@naver.com' || username === 're3@naver.com') {
+                location.href = 'http://devus.localhost:3001' + `?authStore=${encodedJson}`;
+              } else {
+                location.href = loginData?.redirections?.home_url + `?authStore=${encodedJson}`;
+              }
+              deleteCookie('access_token');
+              localStorage.removeItem('auth-store');
+              localStorage.removeItem('app-storage');
             } else {
-              location.href = loginData?.redirections?.home_url + `?authStore=${encodedJson}`;
+              console.error('Invalid authStore format: missing state property');
             }
-
-            deleteCookie('access_token');
-            localStorage.removeItem('auth-store');
-            localStorage.removeItem('app-storage');
-          } else {
-            console.error('Invalid authStore format: missing state property');
+          } catch (error) {
+            console.error('Failed to parse authStore:', error);
           }
-        } catch (error) {
-          console.error('Failed to parse authStore:', error);
+        } else {
+          console.warn('authStore is not available in localStorage');
         }
-      } else {
-        console.warn('authStore is not available in localStorage');
       }
     }
   }, [loginData]);
